+++ cpp/QrCodeGeneratorDemo.cpp
@@ -153,7 +153,7 @@ static void doSegmentDemo() {
 	};
 	qrcodegen::BitBuffer bb;
 	for (int c : kanjiChars)
-		bb.appendBits(c, 13);
+		bb.appendBits(static_cast<std::uint32_t>(c), 13);
 	const QrCode qr5 = QrCode::encodeSegments(
 		{QrSegment(QrSegment::Mode::KANJI, kanjiChars.size(), bb)},
 		QrCode::Ecc::LOW);
